Unity 6.5 multiplatform support (#26)
* Revert "Add dynamic Roslyn path resolving for 6000.5+ (#24)"

This reverts commit 5300c3bfcf91f31faaab47f36e6c5f206aeab744.

* wip

* Update installation instructions for version 1.8.0

+++ UnityEditorPatch/Utilities/FileSystemUtility.cs
@@ -23,9 +23,9 @@ public static class FileSystemUtility
         }
     }
 
-    public static void ReplaceDirectory(string directory, string with)
+    public static void ReplaceDirectory(string? directory, string with)
     {
-        if (!Directory.Exists(with)) return;
+        if (string.IsNullOrEmpty(directory) || !Directory.Exists(with)) return;
 
         Console.WriteLine($"Replacing directory '{directory}' -> '{with}'");
         Directory.Delete(directory, recursive: true);
